+++ CHANGES.md
@@ -10,6 +10,7 @@ Changelog
 - Fixed bug publishing a document to a directory.
 - Fixed bug publishing a document without an extension or type specified.
 - Made 'all' a reserved word that cannot be used as a prefix.
+- Added `doorstop import FILE PREFIX` to import from the export formats.
 
 0.6 (2014/5/15)
 ---------------